Upgrade Touki to 0.6 and enable TOUKI0041
