+++ .editorconfig
@@ -46,37 +46,59 @@ dotnet_style_qualification_for_event = false:error
 dotnet_style_predefined_type_for_locals_parameters_members = true:warning
 dotnet_style_predefined_type_for_member_access = true:warning
 
+# Naming is enforced by TOUKI0041 rather than IDE1006. IDE1006 drops all of its built-in
+# conventions when a custom naming rule is defined and treats const fields as static.
+dotnet_diagnostic.IDE1006.severity = none
+
+# TOUKI0041 ships disabled. This setting enables it and governs the severity of every naming rule.
+dotnet_diagnostic.TOUKI0041.severity = error
+
 # name all constant fields using PascalCase
-dotnet_naming_rule.constant_fields_should_be_pascal_case.severity = suggestion
-dotnet_naming_rule.constant_fields_should_be_pascal_case.symbols  = constant_fields
-dotnet_naming_rule.constant_fields_should_be_pascal_case.style = pascal_case_style
+touki_naming_rule.constant_fields_should_be_pascal_case.severity = suggestion
+touki_naming_rule.constant_fields_should_be_pascal_case.symbols = constant_fields
+touki_naming_rule.constant_fields_should_be_pascal_case.style = pascal_case_style
 
-dotnet_naming_symbols.constant_fields.applicable_kinds   = field
-dotnet_naming_symbols.constant_fields.required_modifiers = const
+touki_naming_symbols.constant_fields.applicable_kinds = field
+touki_naming_symbols.constant_fields.required_modifiers = const
 
-dotnet_naming_style.pascal_case_style.capitalization = pascal_case
+touki_naming_style.pascal_case_style.capitalization = pascal_case
+
+# thread-static fields should have a t_ prefix
+touki_naming_rule.thread_static_fields_should_have_prefix.severity = suggestion
+touki_naming_rule.thread_static_fields_should_have_prefix.symbols = thread_static_fields
+touki_naming_rule.thread_static_fields_should_have_prefix.style = thread_static_prefix_style
+
+touki_naming_symbols.thread_static_fields.applicable_kinds = field
+touki_naming_symbols.thread_static_fields.required_modifiers = static
+touki_naming_symbols.thread_static_fields.required_attributes = System.ThreadStaticAttribute
+
+touki_naming_style.thread_static_prefix_style.required_prefix = t_
+touki_naming_style.thread_static_prefix_style.capitalization = camel_case
 
 # static fields should have s_ prefix
-dotnet_naming_rule.static_fields_should_have_prefix.severity = suggestion
-dotnet_naming_rule.static_fields_should_have_prefix.symbols  = static_fields
-dotnet_naming_rule.static_fields_should_have_prefix.style = static_prefix_style
+touki_naming_rule.static_fields_should_have_prefix.severity = suggestion
+touki_naming_rule.static_fields_should_have_prefix.symbols = static_fields
+touki_naming_rule.static_fields_should_have_prefix.style = static_prefix_style
 
-dotnet_naming_symbols.static_fields.applicable_kinds   = field
-dotnet_naming_symbols.static_fields.required_modifiers = static
+touki_naming_symbols.static_fields.applicable_kinds = field
+touki_naming_symbols.static_fields.required_modifiers = static
 
-dotnet_naming_style.static_prefix_style.required_prefix = s_
-dotnet_naming_style.static_prefix_style.capitalization = camel_case
+touki_naming_style.static_prefix_style.required_prefix = s_
+touki_naming_style.static_prefix_style.capitalization = camel_case
 
 # internal and private fields should be _camelCase
-dotnet_naming_rule.camel_case_for_private_internal_fields.severity = suggestion
-dotnet_naming_rule.camel_case_for_private_internal_fields.symbols  = private_internal_fields
-dotnet_naming_rule.camel_case_for_private_internal_fields.style = camel_case_underscore_style
+touki_naming_rule.camel_case_for_private_internal_fields.severity = suggestion
+touki_naming_rule.camel_case_for_private_internal_fields.symbols = private_internal_fields
+touki_naming_rule.camel_case_for_private_internal_fields.style = camel_case_underscore_style
 
-dotnet_naming_symbols.private_internal_fields.applicable_kinds = field
-dotnet_naming_symbols.private_internal_fields.applicable_accessibilities = private, internal
+touki_naming_symbols.private_internal_fields.applicable_kinds = field
+touki_naming_symbols.private_internal_fields.applicable_accessibilities = private, internal
 
-dotnet_naming_style.camel_case_underscore_style.required_prefix = _
-dotnet_naming_style.camel_case_underscore_style.capitalization = camel_case
+touki_naming_style.camel_case_underscore_style.required_prefix = _
+touki_naming_style.camel_case_underscore_style.capitalization = camel_case
+
+# Keep compile-time-sized stack allocations within the repository's 520-byte ceiling.
+dotnet_code_quality.TOUKI0011.max_stackalloc_bytes = 520
 
 # Code style defaults
 dotnet_sort_system_directives_first = true
@@ -384,10 +406,6 @@ dotnet_diagnostic.IDE0230.severity = suggestion
 
 # IDE1005: Delegate invocation can be simplified.
 dotnet_diagnostic.IDE1005.severity = warning
-
-# IDE1006: Naming styles
-# This rule is configured by individual rules in .editorconfig. Do not override the severity here.
-#dotnet_diagnostic.IDE1006.severity = silent
 
 # IDE2000: Allow multiple blank lines
 dotnet_diagnostic.IDE2000.severity = silent
@@ -2034,9 +2052,6 @@ dotnet_diagnostic.IDE0270.severity = warning
 
 # IDE0300: Simplify collection initialization
 dotnet_diagnostic.IDE0300.severity = warning
-
-# IDE1006: Naming Styles
-dotnet_diagnostic.IDE1006.severity = error
 
 # CA1070: Do not declare event fields as virtual
 dotnet_diagnostic.CA1070.severity = error
