+++ madowaku/Windows/Win32/Foundation/HGLOBAL.cs
@@ -28,57 +28,4 @@ public unsafe partial struct HGLOBAL
     ///  Returns a value indicating whether the handle is null or invalid.
     /// </summary>
     public bool IsValid => !IsNull && PInvoke.GlobalFlags(this) != PInvoke.GMEM_INVALID_HANDLE;
-
-    /// <summary>
-    ///  Locks the global memory handle and returns a pointer to the memory.
-    /// </summary>
-    /// <remarks>
-    ///  <para>
-    ///   Use in a <see langword="using"/> statement to ensure the memory is unlocked when done.
-    ///  </para>
-    /// </remarks>
-    public unsafe ref struct LockScope
-    {
-        private readonly HGLOBAL _global;
-        private void* _memory;
-
-        /// <summary>
-        ///  Constructs a new <see cref="LockScope"/> for the specified global memory handle.
-        /// </summary>
-        public LockScope(HGLOBAL global)
-        {
-            _global = global;
-            _memory = PInvoke.GlobalLock(global);
-            if (_memory is null)
-            {
-                Error.ThrowLastError();
-            }
-        }
-
-        /// <summary>
-        ///  The pointer to the locked memory.
-        /// </summary>
-        public readonly void* Memory => _memory;
-
-        /// <summary>
-        ///  Implicit converter to the pointer to the locked memory.
-        /// </summary>
-        public static implicit operator void*(LockScope scope) => scope.Memory;
-
-        /// <inheritdoc cref="IDisposable.Dispose"/>
-        public void Dispose()
-        {
-            if (_memory is null)
-            {
-                return;
-            }
-
-            if (!PInvoke.GlobalUnlock(_global))
-            {
-                Error.ThrowIfLastErrorNot(WIN32_ERROR.ERROR_SUCCESS);
-            }
-
-            _memory = null;
-        }
-    }
 }
