+++ madowaku/Windows/Win32/System/Com/ComScope{T}.cs
@@ -32,6 +32,7 @@ namespace Windows.Win32.System.Com;
 /// <typeparam name="T">
 ///  This should be one of the struct COM definitions as generated by CsWin32.
 /// </typeparam>
+#pragma warning disable TOUKI0021 // Generic parameters are intentionally included in the file name; Touki #248.
 public readonly unsafe ref struct ComScope<T> where T : unmanaged, IComIID
 {
     // Keeping internal as nint allows us to use Unsafe methods to get significantly better generated code.
@@ -120,3 +121,4 @@ public readonly unsafe ref struct ComScope<T> where T : unmanaged, IComIID
         }
     }
 }
+#pragma warning restore TOUKI0021
